+++ p5-analysis/src/index.ts
@@ -1,3 +1,3 @@
 export { Library } from './Library';
-export { Sketch, SketchType, DirectoryExistsError } from './Sketch';
+export { Sketch, SketchType } from './Sketch';
 export { Script, JavaScriptSyntaxError } from './Script';