+++ p5-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [Unreleased]
+
+- Change options to `p5 create`
+
 ## [0.3.5]
 
 - Server.close is async